Fixed skipp frames on client

+++ src/client.py
@@ -7,6 +7,7 @@ import time
 from os import path
 from pathlib import Path
 from socket import *
+from typing import Dict
 
 from objs import Frame
 from objs.ack import Ack
@@ -21,6 +22,10 @@ PRIORITY_THRESHOLD: Frame.Priority = Frame.Priority(
     int(config["DEFAULT"]["PriorityThreshold"])
 )
 CACHE_PATH: str = config["CLIENT"]["CachePath"]
+FILE_WAIT_TIME: float = 0.005
+FILE_MAX_WAIT_TIME: float = 0.02
+
+frames: Dict[int, FrameBuilder] = {}
 
 
 def writer(client_socket, meta_data: Metadata):
@@ -30,8 +35,6 @@ def writer(client_socket, meta_data: Metadata):
             meta_data.file_name, meta_data.number_of_frames
         )
     )
-
-    frames = {}
     completed_frames = 0
     while completed_frames < meta_data.number_of_frames:
         msg_from = client_socket.recv(1024)
@@ -42,7 +45,7 @@ def writer(client_socket, meta_data: Metadata):
         # check if frame # of packet is in frames here
 
         if p.frame_no not in frames:
-            frames[p.frame_no] = FrameBuilder(p.total_seq_no)
+            frames[p.frame_no] = FrameBuilder(n_expected_packets=p.total_seq_no, priority=p.priority)
         try:
             frames[p.frame_no].emplace(p.seq_no, p.data)
         except Exception:
@@ -64,18 +67,27 @@ def writer(client_socket, meta_data: Metadata):
 def reader(meta_data: Metadata):
     logging.info("Reader Started")
     frame_no = 1
+    while not path.exists("{}{}.h264".format(CACHE_PATH, 1)):
+        time.sleep(FILE_WAIT_TIME)
     while frame_no < meta_data.number_of_frames:
         logging.info("Waiting for {}{}.h264 exists".format(CACHE_PATH, frame_no))
-        while not path.exists("{}{}.h264".format(CACHE_PATH, frame_no)):
-            time.sleep(1)  # force context switch
-        logging.info("Starting {}{}.h264".format(CACHE_PATH, frame_no))
+        time_passed = 0
+        while not path.exists("{}{}.h264".format(CACHE_PATH, frame_no)) and (time_passed < FILE_MAX_WAIT_TIME or (frame_no in frames and frames[frame_no].priority >= PRIORITY_THRESHOLD)):
+            time_passed += FILE_WAIT_TIME
+            time.sleep(FILE_WAIT_TIME)  # force context switch
+
+        if not path.exists(
+                "{}{}.h264".format(CACHE_PATH, frame_no)
+        ):  # skip if frame does not exist
+            logging.warning("Skipping Frame {}".format(frame_no))
+            frame_no += 1
+            continue
 
         with open("{}{}.h264".format(CACHE_PATH, frame_no), "rb") as f:
             with os.fdopen(sys.stdout.fileno(), "wb", closefd=False) as stdout:
                 stdout.write(f.read())
                 stdout.flush()
-        logging.info("Finished {}{}.h264".format(CACHE_PATH, frame_no))
-        logging.info("Deleting {}{}.h264".format(CACHE_PATH, frame_no))
+        logging.info("Wrote {}{}.h264".format(CACHE_PATH, frame_no))
         os.remove("{}{}.h264".format(CACHE_PATH, frame_no))
         frame_no += 1
     logging.info("Reader Finished")
